cc_user_num -> cc_club_num으로 컬럼명 변경에 따라 vo 변경 및 클럽 생성 삭제에 따라 CC_club_num 갯수 변경

+++ sagol/src/main/java/com/sagol/svc/clubSvcImpl.java
@@ -11,8 +11,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.sagol.dao.ccDao;
 import com.sagol.dao.clubDao;
 import com.sagol.dao.clubmemDao;
+import com.sagol.dto.ccVO;
 import com.sagol.dto.clubVO;
 import com.sagol.dto.clubmemVO;
 import com.sagol.dto.searchVO;
@@ -28,6 +30,9 @@ public class clubSvcImpl implements clubSvc {
 	
 	@Autowired
 	private clubmemDao clubmemdao;
+	
+	@Autowired
+	private ccDao ccdao;
 
 
 	@Override
@@ -61,7 +66,7 @@ public class clubSvcImpl implements clubSvc {
 		int result = clubdao.insertClub(clubvo);
 		// 클럽생성
 		
-		if (result == 1) {//클럽생성 성공시 생성자의 세션 정보로 클럽멤버 추가
+		if (result == 1) {
 			clubmemVO clubmemvo = new clubmemVO();
 			HttpSession session = request.getSession();
 			userVO uservo = (userVO) session.getAttribute("userVO");
@@ -70,6 +75,11 @@ public class clubSvcImpl implements clubSvc {
 			clubmemvo.setOwner_yn("Y");
 			clubmemvo.setReg_dt(time);
 			clubmemdao.insertClubmem(clubmemvo);
+			//클럽생성 성공시 생성자의 세션 정보로 클럽멤버 추가
+			ccVO ccvo = new ccVO();
+			ccvo.setCc_id(clubvo.getCc_id());
+			ccdao.addClubNum(ccvo);
+			//소속 CC의 클럽수를 1 증가한다.
 		}//@Transactional로 인해 실패시 올 롤백됨
 		
 		return result;
@@ -85,8 +95,8 @@ public class clubSvcImpl implements clubSvc {
 		clubmemvo.setClub_id(clubvo.getClub_id());
 		
 		clubmemVO resultclubmemvo = clubmemdao.selectOwnerByClubId(clubmemvo);
-		System.out.println(!(sessionUservo.getUid().equals(resultclubmemvo.getUid())));
-		System.out.println(!("Y".equals(sessionUservo.getAdmin_yn())));
+//		System.out.println(!(sessionUservo.getUid().equals(resultclubmemvo.getUid())));
+//		System.out.println(!("Y".equals(sessionUservo.getAdmin_yn())));
 		if (!(sessionUservo.getUid().equals(resultclubmemvo.getUid())) && !("Y".equals(sessionUservo.getAdmin_yn()))) {
 			//업데이트 행위자가 owner가 아니거나 어드민이 아닐경우
 			return 3;
@@ -101,15 +111,36 @@ public class clubSvcImpl implements clubSvc {
 	}
 
 	@Override
-	public int deleteClub(clubVO clubvo) {
+	public int deleteClub(clubVO clubvo,HttpServletRequest request) {
+		
+		HttpSession session = request.getSession();
+		userVO sessionUservo = (userVO) session.getAttribute("userVO");	
+		
+		clubmemVO clubmemvo = new clubmemVO();
+		clubmemvo.setClub_id(clubvo.getClub_id());
+		
+		clubmemVO resultclubmemvo = clubmemdao.selectOwnerByClubId(clubmemvo);
+//		System.out.println(!(sessionUservo.getUid().equals(resultclubmemvo.getUid())));
+//		System.out.println(!("Y".equals(sessionUservo.getAdmin_yn())));
+		if (!(sessionUservo.getUid().equals(resultclubmemvo.getUid())) && !("Y".equals(sessionUservo.getAdmin_yn()))) {
+			//업데이트 행위자가 owner가 아니거나 어드민이 아닐경우
+			return 3;
+		}
 		if (clubdao.isExistByClubId(clubvo) == 0) {
 			return 2;
 		}
-		clubmemVO clubmemvo = new clubmemVO();
-		clubmemvo.setClub_id(clubvo.getClub_id());
+//		clubmemVO clubmemvo = new clubmemVO();
+//		clubmemvo.setClub_id(clubvo.getClub_id());
 		clubmemdao.deleteClubmemsByClubid(clubmemvo);
 		//외래키 제약으로 인해 Club 하위 멤버들 먼저 삭제 후 Club삭제 진행
-		return clubdao.deleteClub(clubvo);
+		int result = clubdao.deleteClub(clubvo);
+		if (result ==1) {
+			ccVO ccvo = new ccVO();
+			ccvo.setCc_id(clubvo.getCc_id());
+			ccdao.minusClubNum(ccvo);
+			//클럽 삭제가 성공되면 CC의 클럽수를 1 줄인다.
+		}
+		return result;
 
 	}
